# src/main/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.2.3-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `https://MY-REGISTRY-URL/apis/registry/v2` by default. Therefore you must prefix all API operation paths with `../apis/registry/v2` in this case. For example: `../apis/registry/v2/ids/globalIds/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
...
    },
    "paths": {
...
        "/groups/{groupId}/artifacts/{artifactId}/state": {
            "summary": "Manage the state of an artifact.",
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateState"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Artifacts"
                ],
                "responses": {
                    "204": {
                        "description": "Returned when the operation was successful."
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateArtifactState",
                "summary": "Update artifact state",
                "description": "Updates the state of the artifact.  For example, you can use this to mark the latest\nversion of an artifact as `DEPRECATED`.  The operation changes the state of the latest \nversion of the artifact.  If multiple versions exist, only the most recent is changed.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "groupId",
                    "description": "The artifact group ID.  Must be a string provided by the client, representing the name of the grouping of artifacts.",
                    "schema": {
                        "$ref": "#/components/schemas/GroupId"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "artifactId",
                    "description": "The artifact ID.  Can be a string (client-provided) or UUID (server-generated), representing the unique artifact identifier.",
                    "schema": {
                        "$ref": "#/components/schemas/ArtifactId"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/groups/{groupId}/artifacts/{artifactId}/versions/{version}/meta": {
            "summary": "Manage the metadata for a single version of an artifact in the registry.",
            "get": {
                "tags": [
                    "Metadata"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionMetaData"
                                }
                            }
                        },
                        "description": "The artifact version's metadata."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getArtifactVersionMetaData",
                "summary": "Get artifact version metadata",
                "description": "Retrieves the metadata for a single version of the artifact.  The version metadata is \na subset of the artifact metadata and only includes the metadata that is specific to\nthe version (for example, this doesn't include `modifiedOn`).\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No version with this `version` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditableMetaData"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Metadata"
                ],
                "responses": {
                    "204": {
                        "description": "The artifact version's metadata was successfully updated."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateArtifactVersionMetaData",
                "summary": "Update artifact version metadata",
                "description": "Updates the user-editable portion of the artifact version's metadata.  Only some of \nthe metadata fields are editable by the user.  For example, `description` is editable, \nbut `createdOn` is not.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No version with this `version` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Metadata"
                ],
                "responses": {
                    "204": {
                        "description": "The artifact version's user-editable metadata was successfully deleted."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteArtifactVersionMetaData",
                "summary": "Delete artifact version metadata",
                "description": "Deletes the user-editable metadata properties of the artifact version.  Any properties\nthat are not user-editable are preserved.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No version with this `version` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "groupId",
                    "description": "The artifact group ID.  Must be a string provided by the client, representing the name of the grouping of artifacts.",
                    "schema": {
                        "$ref": "#/components/schemas/GroupId"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "artifactId",
                    "description": "The artifact ID.  Can be a string (client-provided) or UUID (server-generated), representing the unique artifact identifier.",
                    "schema": {
                        "$ref": "#/components/schemas/ArtifactId"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "version",
                    "description": "The unique identifier of a specific version of the artifact content.",
                    "schema": {
                        "$ref": "#/components/schemas/Version"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateState": {
                "title": "Root Type for UpdateState",
                "description": "",
                "required": [
                    "state"
                ],
                "type": "object",
                "properties": {
                    "state": {
                        "$ref": "#/components/schemas/ArtifactState"
                    }
                },
                "example": {
                    "state": "DISABLED"
                }
            },
...
        },
        "responses": {
            "NotFound": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "NotFoundExample": {
                                "value": {
                                    "error_code": 404,
                                    "message": "No artifact with id 'Topic-1/Inbound' could be found."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can return a `404` error."
            },
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
            "BadRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                },
                "description": "Common response for all operations that can return a `400` error."
            },
...
        }
    },
...
}